adding logic for respawn with invicibility
respawn check for invincibility

+++ js/invaders.js
@@ -71,6 +71,8 @@ function create() {
     player.anchor.setTo(0.5, 0.5);
     // disable texture smoothing to make pixel art stand out
     player.smoothed = false;
+    // set the custom invincible property to false by default. Used when respawning for temporary invinciblity.
+    player.invincible = false;
     game.physics.enable(player, Phaser.Physics.ARCADE);
 
 
@@ -238,7 +240,7 @@ function update() {
         enemyFires();
     }
 
-    //  Run collision
+    //  Run collision detection
     game.physics.arcade.overlap(bullets, aliens, collisionHandler, null, this);
     game.physics.arcade.overlap(enemyBullets, player, enemyHitsPlayer, null, this);
 
@@ -286,46 +288,53 @@ function collisionHandler (bullet, alien) {
 
 function enemyHitsPlayer (player,bullet) {
     
-    bullet.kill();
-
-    live = lives.getFirstAlive();
-
-    if (live)
+    if (player.invincible = false)
     {
-        live.kill();
-    }
+	    bullet.kill();
+
+	    live = lives.getFirstAlive();
+
+	    if (live)
+	    {
+	        live.kill();
+	    }
 
 
-    //  And create an explosion :)
-    var explosion = explosions.getFirstExists(false);
-    explosion.reset(player.body.x, player.body.y);
-    explosion.play('kaboom', 30, false, true);               
+	    //  And create an explosion :)
+	    var explosion = explosions.getFirstExists(false);
+	    explosion.reset(player.body.x, player.body.y);
+	    explosion.play('kaboom', 30, false, true);               
 
-    // When the player dies
-    if (lives.countLiving() < 1)
-    {
-        player.kill();
-        enemyBullets.callAll('kill');
-        gameisrunning = 0;
+	    // When the player dies
+	    if (lives.countLiving() < 1)
+	    {
+	        player.kill();
+	        enemyBullets.callAll('kill');
+	        gameisrunning = 0;
 
-        //font.text = 'GAME OVER'
+	        //font.text = 'GAME OVER'
 
-        //stateText.text=" GAME OVER \n Click to restart";
-        //stateText.visible = true;
+	        //stateText.text=" GAME OVER \n Click to restart";
+	        //stateText.visible = true;
 
-        //the "click to restart" handler
-        game.input.onTap.addOnce(restart,this);        
-    }
-    else
-    {
-        player.alpha = 0.4;
-        game.time.events.add(Phaser.Timer.SECOND * 3, function(){player.alpha = 1.0;})    
-        player.body.x = 320;
-        player.body.y = 800; 
+	        //the "click to restart" handler
+	        game.input.onTap.addOnce(restart,this);        
+	    }
+	    else
+	    {
+	        // set the player to slightly transparent
+	        player.alpha = 0.4;
+	        // player is invincible
+	        player.invincible = true;
+	        game.time.events.add(Phaser.Timer.SECOND * 3, function(){player.alpha = 1.0; player.invincible = false;})          
+	        player.body.x = 320;
+	        player.body.y = 800; 
 
-    }
-
+	    }
+	}
 }
+
+
 
 function enemyFires () {
